fix: golangci-lint v1 config schema and ARMv7 Bsize type cast
- Remove version/linters.default keys invalid in golangci-lint v1.x
- Cast s.Bsize to int64 in statfs_linux.go — Statfs_t.Bsize is int32
  on ARMv7 but int64 on arm64/amd64, causing cross-compile failure

+++ internal/watchers/statfs_linux.go
@@ -16,7 +16,7 @@ func statfsLinux(path string, stat *StatFS) error {
 	if err := syscall.Statfs(path, &s); err != nil {
 		return err
 	}
-	stat.Bsize = s.Bsize
+	stat.Bsize = int64(s.Bsize) // Bsize is int32 on ARMv7, int64 on arm64/amd64
 	stat.Blocks = s.Blocks
 	stat.Bfree = s.Bfree
 	return nil
